requirements.txt
asgiref==3.3.4
attrs==21.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
djangify==1.0.0
Django==3.2.3
django-allauth==0.44.0
django-bootstrap4==3.0.1
django-crispy-forms==1.11.2
django-filter==2.4.0
django-rest-swagger==2.2.0
django-simple-history==3.0.0
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jsonschema==3.2.0
Markdown==3.3.4
MarkupSafe==2.0.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==21.0
Pillow==8.2.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
pyrsistent==0.18.0
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
simplejson==3.17.2
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
swagger-spec-validator==2.7.3
uritemplate==3.0.1
urllib3==1.26.5
